GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - gitbutler/integration (refs/gitbutler/gitbutler/integration)
   branch head: 3b462bf1959484338eec31ba5a396e459373dff6

Your previous branch was: refs/heads/main

The sha for that commit was: 7243d3d46ffcb274a79774aaacc97303131b0464

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch