feat: Implement adding a new person to the repository
Added CreatePersonInputDto class with FirstName, LastName, and PhoneNumber properties.
Implemented AddPerson method in PersonApplicationService to create a new Person object and add it to the repository.
Added IPersonRepository interface with Add method.
Updated Person constructor to accept a single PhoneNumber instead of a list.
Added Ms.Core.ApplicationServices project to the solution file.